separated integration test
